# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat Atendimento API",
    "version": "1.0.0",
    "description": "\n        API REST para sistema de chat de atendimento em tempo real.\n        \n        ## Funcionalidades\n        - Autenticação JWT\n        - Chat em tempo real com WebSocket\n        - Upload de arquivos (AWS S3)\n        - Gerenciamento de atendentes\n        - Fila de atendimento\n        - Dashboard administrativo\n        - Histórico e análises\n        \n        ## Ambientes\n        - **Desenvolvimento**: http://localhost:5000\n        - **Staging**: http://52.90.17.204:5000\n        - **Produção**: https://api.brsi.com.br\n        \n        ## Autenticação\n        Todas as rotas (exceto login e register) requerem autenticação via JWT.\n        Envie o token no header: Authorization: Bearer {token}\n      ",
    "contact": {
      "name": "BR Sistemas",
      "email": "suporte@brsi.com.br",
      "url": "https://brsi.com.br"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://brsi.com.br/terms"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desenvolvimento (Local)"
    },
    {
      "url": "http://52.90.17.204:5000/api",
      "description": "Servidor de Staging (Testes)"
    },
    {
      "url": "https://api.brsi.com.br/api",
      "description": "Servidor de Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Digite o token JWT obtido no login"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único do usuário",
            "example": "60d0fe4f5311236168a109ca"
          },
          "name": {
            "type": "string",
            "description": "Nome completo do usuário",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário",
            "example": "joao@brsi.com.br"
          },
          "role": {
            "type": "string",
            "enum": [
              "agent",
              "admin",
              "client"
            ],
            "description": "Papel do usuário no sistema",
            "example": "agent"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "away",
              "busy",
              "offline"
            ],
            "description": "Status atual do usuário",
            "example": "online"
          },
          "isActive": {
            "type": "boolean",
            "description": "Se o usuário está ativo",
            "example": true
          },
          "avatar": {
            "type": "string",
            "description": "URL do avatar do usuário",
            "example": "https://chat-staging.s3.amazonaws.com/avatars/user123.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Última atualização",
            "example": "2024-01-20T15:45:00Z"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único da conversa",
            "example": "60d0fe4f5311236168a109cb"
          },
          "subject": {
            "type": "string",
            "description": "Assunto da conversa",
            "example": "Dúvida sobre produto"
          },
          "customer": {
            "type": "object",
            "description": "Dados do cliente",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "example": "Maria Santos"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "maria@example.com"
              }
            }
          },
          "agent": {
            "type": "string",
            "description": "ID do agente atribuído",
            "example": "60d0fe4f5311236168a109ca"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "closed"
            ],
            "description": "Status atual da conversa",
            "example": "active"
          },
          "department": {
            "type": "string",
            "description": "Departamento responsável",
            "example": "vendas"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Avaliação do atendimento",
            "example": 5
          },
          "ratingComment": {
            "type": "string",
            "description": "Comentário da avaliação",
            "example": "Atendimento excelente!"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de encerramento"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "Última atividade na conversa"
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/agents/profile/{agentId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obter perfil do agente",
        "description": "Retorna o perfil de um agente específico ou do agente atual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ID do agente (opcional, usa o usuário atual se não fornecido)"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil do agente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "online",
                        "away",
                        "busy",
                        "offline"
                      ]
                    },
                    "availability": {
                      "type": "boolean"
                    },
                    "currentLoad": {
                      "type": "integer"
                    },
                    "maxLoad": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Agente não encontrado"
          }
        }
      }
    },
    "/api/agents/status": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Atualizar status do agente",
        "description": "Atualiza o status do agente (online, away, busy, offline)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "online",
                      "away",
                      "busy",
                      "offline"
                    ],
                    "example": "online"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "agent": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Status inválido"
          },
          "403": {
            "description": "Sem permissão - Apenas agentes e administradores"
          }
        }
      }
    },
    "/api/agents/availability": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Atualizar disponibilidade",
        "description": "Define se o agente está disponível para receber novos atendimentos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "available"
                ],
                "properties": {
                  "available": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disponibilidade atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "availability": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão"
          }
        }
      }
    },
    "/api/agents/preferences": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Atualizar preferências do agente",
        "description": "Atualiza preferências de atendimento do agente",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxConcurrentChats": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "example": 3
                  },
                  "notificationsEnabled": {
                    "type": "boolean",
                    "example": true
                  },
                  "soundEnabled": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferências atualizadas"
          },
          "403": {
            "description": "Sem permissão"
          }
        }
      }
    },
    "/api/agents/performance": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obter performance do agente",
        "description": "Retorna métricas de performance do agente atual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month"
              ],
              "default": "today"
            },
            "description": "Período para análise"
          }
        ],
        "responses": {
          "200": {
            "description": "Métricas de performance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalChats": {
                      "type": "integer"
                    },
                    "resolvedChats": {
                      "type": "integer"
                    },
                    "avgResponseTime": {
                      "type": "number"
                    },
                    "avgResolutionTime": {
                      "type": "number"
                    },
                    "satisfaction": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 5
                    },
                    "totalRatings": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/performance/{agentId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obter performance de agente específico",
        "description": "Retorna métricas de performance de um agente específico (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do agente"
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month"
              ],
              "default": "today"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Métricas de performance"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Agente não encontrado"
          }
        }
      }
    },
    "/api/agents/workload": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obter distribuição de carga de trabalho",
        "description": "Retorna a distribuição atual de conversas entre agentes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Distribuição de carga de trabalho",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentId": {
                        "type": "string"
                      },
                      "agentName": {
                        "type": "string"
                      },
                      "currentLoad": {
                        "type": "integer"
                      },
                      "maxLoad": {
                        "type": "integer"
                      },
                      "status": {
                        "type": "string"
                      },
                      "activeChats": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão"
          }
        }
      }
    },
    "/api/agents/all": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Listar todos os agentes",
        "description": "Retorna lista de todos os agentes do sistema",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "away",
                "busy",
                "offline"
              ]
            },
            "description": "Filtrar por status"
          },
          {
            "in": "query",
            "name": "available",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por disponibilidade"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de agentes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão"
          }
        }
      }
    },
    "/api/agents/auto-assign": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Auto-atribuir conversa",
        "description": "Atribui automaticamente uma conversa ao agente mais adequado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conversationId"
                ],
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109ca"
                  },
                  "department": {
                    "type": "string",
                    "example": "vendas"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversa atribuída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "assignedAgent": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "conversation": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nenhum agente disponível"
          },
          "403": {
            "description": "Sem permissão"
          }
        }
      }
    },
...
  }
}